Back to float

+++ assignment5/src/main.c
@@ -11,7 +11,7 @@
 #include "read_input.h"
 #include "diffusion.h"
 
-void printdata(double* data, int rows, int cols, int mpi_rank) {
+void printdata(float* data, int rows, int cols, int mpi_rank) {
     printf("Rank %d: \n", mpi_rank);
     for (int i = 0; i < rows + 2; ++i) {
         for (int j = 0; j < cols + 2; ++j) {
@@ -36,44 +36,44 @@ main(int argc, char **argv)
 
     // Read data
     size_t rows, cols, global_rows;
-    double *dataA = read_data(mpi_rank, nmb_mpi_proc, &rows, &cols, &global_rows);
-    double *dataB = calloc((rows + 2)*(cols+2), sizeof(double));
+    float *dataA = read_data(mpi_rank, nmb_mpi_proc, &rows, &cols, &global_rows);
+    float *dataB = calloc((rows + 2)*(cols+2), sizeof(float));
 
     if (args.verbose && mpi_rank == 0)
         printf("Data read\n");
 
     for ( size_t step = 0; step < args.n_iter; step++) {
-        double *data = step % 2 ? dataB : dataA;
-        double *next_data = step % 2 ? dataA : dataB;
+        float *data = step % 2 ? dataB : dataA;
+        float *next_data = step % 2 ? dataA : dataB;
 
         if (args.verbose && mpi_rank == 0)
             printf("In step %lu\n", step);
 
         // Sync BC
-        double *padding_up = data + 1;
-        double *padding_down = data + (cols + 2) * (rows + 1) + 1;
-        double *data_first = data + cols + 2 + 1;
-        double *data_last = padding_down - cols - 2;
+        float *padding_up = data + 1;
+        float *padding_down = data + (cols + 2) * (rows + 1) + 1;
+        float *data_first = data + cols + 2 + 1;
+        float *data_last = padding_down - cols - 2;
 
         MPI_Status status;
         if (nmb_mpi_proc > 1) {
             if ( mpi_rank == 0) {
-                MPI_Sendrecv(data_last, cols, MPI_DOUBLE, 1, 0,
-                            padding_down, cols, MPI_DOUBLE, 1, 0,
+                MPI_Sendrecv(data_last, cols, MPI_FLOAT, 1, 0,
+                            padding_down, cols, MPI_FLOAT, 1, 0,
                             MPI_COMM_WORLD, &status);
             }
             else if ( mpi_rank == nmb_mpi_proc - 1) {
-                MPI_Sendrecv(data_first, cols, MPI_DOUBLE, mpi_rank - 1, 0,
-                            padding_up, cols, MPI_DOUBLE, mpi_rank - 1, 0,
+                MPI_Sendrecv(data_first, cols, MPI_FLOAT, mpi_rank - 1, 0,
+                            padding_up, cols, MPI_FLOAT, mpi_rank - 1, 0,
                             MPI_COMM_WORLD, &status);
             }
             else {
-                MPI_Sendrecv(data_first, cols, MPI_DOUBLE, mpi_rank - 1, 0,
-                            padding_down, cols, MPI_DOUBLE, mpi_rank + 1, 0,
+                MPI_Sendrecv(data_first, cols, MPI_FLOAT, mpi_rank - 1, 0,
+                            padding_down, cols, MPI_FLOAT, mpi_rank + 1, 0,
                             MPI_COMM_WORLD, &status);
 
-                MPI_Sendrecv(data_last, cols, MPI_DOUBLE, mpi_rank + 1, 0,
-                            padding_up, cols, MPI_DOUBLE, mpi_rank - 1, 0,
+                MPI_Sendrecv(data_last, cols, MPI_FLOAT, mpi_rank + 1, 0,
+                            padding_up, cols, MPI_FLOAT, mpi_rank - 1, 0,
                             MPI_COMM_WORLD, &status);
             }
         }
@@ -86,10 +86,10 @@ main(int argc, char **argv)
     if (args.verbose && mpi_rank == 0)
             printf("Loop done\n");
 
-    double *data = args.n_iter % 2 ? dataB : dataA;
+    float *data = args.n_iter % 2 ? dataB : dataA;
 
     // Calculate average
-    double internal_avg = 0.0f;
+    float internal_avg = 0.0f;
     for ( size_t i = 1; i <= rows; i ++) {
         for ( int j = 1; j <= cols; j++) {
             int index = i * (cols + 2) + j;
@@ -99,14 +99,14 @@ main(int argc, char **argv)
     internal_avg /= global_rows * cols;
 
     // Reduce average to all processes
-    double avg;
-    MPI_Allreduce(&internal_avg, &avg, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
+    float avg;
+    MPI_Allreduce(&internal_avg, &avg, 1, MPI_FLOAT, MPI_SUM, MPI_COMM_WORLD);
 
     if (args.verbose && mpi_rank == 0)
         printf("Average calculated to %f\n", avg);
 
     // Calculate abs diff
-    double abs_diff;
+    float abs_diff;
     for ( size_t i = 1; i <= rows; i ++) {
         for ( int j = 1; j <= cols; j++) {
             int index = i * (cols + 2) + j;
@@ -116,8 +116,8 @@ main(int argc, char **argv)
     abs_diff /= global_rows * cols;
 
     // Reduce to mpi rank 0
-    double reduce_abs_diff;
-    MPI_Reduce(&abs_diff, &reduce_abs_diff, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
+    float reduce_abs_diff;
+    MPI_Reduce(&abs_diff, &reduce_abs_diff, 1, MPI_FLOAT, MPI_SUM, 0, MPI_COMM_WORLD);
 
     if (args.verbose && mpi_rank == 0)
         printf("Abs diff Average calculated to %f\n", reduce_abs_diff);
@@ -128,8 +128,8 @@ main(int argc, char **argv)
     free(dataB);
 
     if (mpi_rank == 0) {
-        printf("average: %lf\n", avg);
-        printf("average absolute difference: %lf\n", reduce_abs_diff);
+        printf("average: %f\n", avg);
+        printf("average absolute difference: %f\n", reduce_abs_diff);
     }
 
     return 0;
