+++ assignment5/src/read_input.c
@@ -1,11 +1,11 @@
 #include "read_input.h"
 
-double min(double a, double b) {
+float min(float a, float b) {
     if (a < b) return a;
     return b;
 }
 
-double * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols, size_t *global_rows) {
+float * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols, size_t *global_rows) {
     // Create file pointer
     FILE *file = fopen("init", "r");
 
@@ -18,19 +18,19 @@ double * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols, s
     // Read the first row of the which is the size: [rows, cols]
     fscanf(file, "%zu %zu\n", global_rows, cols);
     // printf("box [%zu x %zu]\n", *rows, *cols);
-    
+
     // Calculate the number of rows each worker will process
     int rows_per_worker = *global_rows / nmb_mpi_proc;
     int remainder_rows = *global_rows % nmb_mpi_proc;
-    
+
     int my_rows = (mpi_rank < remainder_rows) ? (rows_per_worker + 1) : rows_per_worker;
     *rows = my_rows;
-    
+
     int row_zero = rows_per_worker * mpi_rank + min(mpi_rank, remainder_rows);
 
-    
+
     // Initialize data matrix
-    double *data = (double *)calloc((my_rows+2) * (*cols + 2), sizeof(double));
+    float *data = (float *)calloc((my_rows+2) * (*cols + 2), sizeof(float));
 
     if (data == NULL) {
 
@@ -41,16 +41,16 @@ double * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols, s
 
     // Initialize coordinates and val
     int x, y;
-    double val;
+    float val;
 
     // Read values for each coordinates
-    while(fscanf(file, "%d %d %lf", &x, &y, &val) != EOF) {
+    while(fscanf(file, "%d %d %f", &x, &y, &val) != EOF) {
         if (x >= row_zero && x < (row_zero + my_rows)){
             // printf(" x y v = %d %d %f\n", x, y, val);
             data[(x - row_zero + 1)*(*cols+2) + (y+1)] = val;
         }
     }
-    
+
     fclose(file);
 
     return data;
@@ -59,7 +59,7 @@ double * read_data(int mpi_rank, int nmb_mpi_proc, size_t *rows, size_t *cols, s
 #ifdef READ_INPUT_MAIN
 int main(int argc, char** argv) {
 
-    double *data;
+    float *data;
     size_t rows, cols;
     int nmb_mpi_proc = 3;
